Fixed Modal Styling and Props
Finished modal! Still need to write testing, but she's done. I did try to justify the content, but the justifyContent in the styling doesn't seem to be doing anything at the moment. As well, I want to look into fixing the position so that the size doesn't change depending on monitor size, but for now I think this is a good start. I might make a class to put the styling in later.